refactor: Move api functionality into its own controller and make the concerning changes
