+++ routes/web.php
@@ -14,6 +14,4 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::controller(CityController::class)->group(function () {
-    Route::get('/', 'index')->name('index');
-});
+Route::get('/',[CityController::class, 'index']);
